app/data_insertion_service/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
sqlalchemy.exc